equity plot screen adjustment fix

+++ equity.py
@@ -50,7 +50,7 @@ class EquityPlot(QWidget):
         self.resize(self.width, self.height)
         # self.setFixedWidth(self.width)
         # self.setFixedHeight(self.height)
-        self.move(self.screenWidth/2. - self.width/2., self.screenHeight/2. - self.height/2.)
+        self.move(int(self.screenWidth/2. - self.width/2.), int(self.screenHeight/2. - self.height/2.))
 
         # scroll = QScrollArea(w)
         # l.addWidget(scroll)
